+++ packages/form-schema/src/fields/checkbox/checkbox.ts
@@ -7,12 +7,5 @@ export function prepareValue(rawValue: unknown) {
 
 export const checkbox: FieldType<CheckboxSchema> = {
 	serializerSingle: ({ value }) => prepareValue(value),
-
-	parser: ({ name, values }) => {
-		const value = values[name];
-
-		return {
-			[name]: prepareValue(value),
-		};
-	},
+	parserSingle: ({ value }) => prepareValue(value),
 };